Require graphql schema
