+++ pom.xml
@@ -207,12 +207,6 @@
             <groupId>io.undertow</groupId>
             <artifactId>undertow-core</artifactId>
             <scope>compile</scope>
-            <exclusions>
-                <exclusion>
-                    <artifactId>jboss-logging</artifactId>
-                    <groupId>org.jboss.logging</groupId>
-                </exclusion>
-            </exclusions>
         </dependency>
 
         <dependency>
