+++ src/main/java/no/ssb/lds/graphql/schemas/GraphQLSchemaBuilder.java
@@ -11,12 +11,10 @@ import graphql.schema.TypeTraverser;
 import graphql.schema.idl.EchoingWiringFactory;
 import graphql.schema.idl.RuntimeWiring;
 import graphql.schema.idl.SchemaGenerator;
-import graphql.schema.idl.SchemaParser;
 import graphql.schema.idl.SchemaPrinter;
 import graphql.schema.idl.TypeDefinitionRegistry;
 import no.ssb.lds.api.persistence.reactivex.RxJsonPersistence;
 import no.ssb.lds.api.search.SearchIndex;
-import no.ssb.lds.core.specification.JsonSchemaBasedSpecification;
 import no.ssb.lds.graphql.directives.DomainDirective;
 import no.ssb.lds.graphql.directives.LinkDirective;
 import no.ssb.lds.graphql.directives.ReverseLinkDirective;
@@ -30,11 +28,7 @@ import no.ssb.lds.graphql.schemas.visitors.TypeReferencerVisitor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.File;
-import java.util.ArrayDeque;
-import java.util.Arrays;
 import java.util.Collection;
-import java.util.Deque;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
@@ -56,45 +50,6 @@ public class GraphQLSchemaBuilder {
         this.persistence = persistence;
         this.searchIndex = searchIndex;
         this.namespace = namespace;
-    }
-
-    /**
-     * Use this to print a graphql schema out of a directory of json files.
-     */
-    public static void main(String... argv) {
-
-        Boolean graphQl = false;
-        Boolean jsonSchema = false;
-        Deque<String> arguments = new ArrayDeque<>(Arrays.asList(argv));
-        String current = arguments.peek();
-        while (!arguments.isEmpty()) {
-            if ("--graphql".equals(current)) {
-                graphQl = true;
-            }
-            if ("--json-schema".equals(current)) {
-                jsonSchema = true;
-            }
-            current = arguments.pop();
-        }
-
-        GraphQLSchemaBuilder schemaBuilder = new GraphQLSchemaBuilder("namespace", new EmptyPersistence(), null);
-
-        GraphQLSchema schema;
-        if (graphQl && !jsonSchema) {
-            log.info("Parsing graphql schema {}", current);
-            TypeDefinitionRegistry definitionRegistry = new SchemaParser().parse(new File(current));
-            schema = schemaBuilder.parseSchema(definitionRegistry);
-        } else if (jsonSchema && !graphQl) {
-            log.info("Parsing json-schemas in {}", current);
-            JsonSchemaBasedSpecification specification = JsonSchemaBasedSpecification.create(current);
-            SpecificationConverter test = new SpecificationConverter();
-            //schema = schemaBuilder.parseSpecification(specification);
-            schema = schemaBuilder.parseSchema(test.convert(specification));
-        } else {
-            log.error("Usage: --graphql file | --json-schema folder");
-            return;
-        }
-        GraphQLSchema graphQL = schemaBuilder.getGraphQL(schema);
     }
 
     private static String printSchema(GraphQLType type) {
