+++ src/main/resources/application-defaults.properties
@@ -14,16 +14,14 @@ persistence.initialization.max-wait-seconds=0
 
 namespace.default=ns
 
-# The owner of the cluster. Typically mapped to a security context used to discriminate tenants
-cluster.owner=my.org
-
-# The name of the cluster. Can be used to separate environments in a deployment where environments share infrastructure
-cluster.name=lds-test
-
-# The unique instance-id to be used by this running process
-cluster.instance-id=01
-
 sagalog.provider=no.ssb.sagalog.memory.MemorySagaLogInitializer
+# The owner of the cluster. Typically mapped to a security context used to discriminate tenants
+sagalog.config.cluster.owner=my.org
+# The name of the cluster. Can be used to separate environments in a deployment where environments share infrastructure
+sagalog.config.cluster.name=lds-test
+# The unique instance-id to be used by this running process
+sagalog.config.cluster.instance-id=01
+
 
 saga.number-of-logs=50
