+++ src/test/resources/application_test.properties
@@ -16,13 +16,14 @@ specification.schema=spec/schemas/contact.json,spec/schemas/provisionagreement.j
 http.request.dump=false
 graphql.enabled=false
 
-graphql.schema=graphql/schema.graphql
+graphql.schema=gqlschema/schema.graphql
 
 graphql.search.enabled=true
 search.index.provider=testSearchIndex
 
 sagalog.provider=no.ssb.sagalog.file.FileSagaLogInitializer
-filesagalog.folder=./target/test-sagalog
+sagalog.config.filesagalog.folder=./target/test-sagalog
+
 saga.number-of-logs=5
 
 saga.recovery.enabled=false
